Fix the buyAmount and sellAmount type from Int32 to Int64

+++ Model/Quote/MarketData.cs
@@ -93,10 +93,10 @@ namespace Model.Quote
         public double SellPrice5;
 
         //外盘
-        public int BuyAmount;
+        public long BuyAmount;
 
         //内盘
-        public int SellAmount;
+        public long SellAmount;
 
         public SuspendFlag SuspendFlag;
